Create lame coefficient inside CoordSysCartesian.
