+++ sympy/vector/__init__.py
@@ -13,5 +13,4 @@ from sympy.vector.functions import (express, matrix_to_vector,
 from sympy.vector.point import Point
 from sympy.vector.orienters import (AxisOrienter, BodyOrienter,
                                     SpaceOrienter, QuaternionOrienter)
-from sympy.vector.lamecoeff import CoeffProvider
 from sympy.vector.operators import Gradient, Divergence, Curl, gradient, curl, divergence
